requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic-settings==2.10.1
httpx==0.28.1
openai==1.101.0
google-cloud-speech==2.33.0
google-generativeai==0.8.5
ringcentral==0.8.5
requests==2.32.3
python-multipart==0.0.20
